requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
cloud-sql-python-connector[pg8000]>=1.11.0
pg8000>=1.31.0
